# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_StorageRootResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StorageRootResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[StorageRootResponse]",
        "type": "object"
      },
...
      "RegisterRootRequest": {
        "description": "Allowlist an institutional path, on the caller's stated authority.\n\n`attestation_note` is required, not optional. The thing being recorded is\nnot that somebody ticked a box but *what they checked* \u2014 which share, whose\nmembers, against which access list \u2014 and nobody can reconstruct that later.",
        "properties": {
          "attestation_note": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Attestation Note",
            "type": "string"
          },
          "id": {
            "maxLength": 64,
            "title": "Id",
            "type": "string"
          },
          "label": {
            "maxLength": 256,
            "title": "Label",
            "type": "string"
          },
          "readable": {
            "default": true,
            "title": "Readable",
            "type": "boolean"
          },
          "root_path": {
            "maxLength": 1024,
            "title": "Root Path",
            "type": "string"
          },
          "writable": {
            "default": false,
            "title": "Writable",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "label",
          "root_path",
          "attestation_note"
        ],
        "title": "RegisterRootRequest",
        "type": "object"
      },
      "ReinstateRootRequest": {
        "properties": {
          "attestation_note": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Attestation Note",
            "type": "string"
          },
          "readable": {
            "default": true,
            "title": "Readable",
            "type": "boolean"
          },
          "writable": {
            "default": false,
            "title": "Writable",
            "type": "boolean"
          }
        },
        "required": [
          "attestation_note"
        ],
        "title": "ReinstateRootRequest",
        "type": "object"
      },
...
      "RevokeRootRequest": {
        "properties": {
          "reason": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "RevokeRootRequest",
        "type": "object"
      },
...
      "StorageRootResponse": {
        "properties": {
          "attestation_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attestation Note"
          },
          "attested_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attested At"
          },
          "attested_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attested By"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "identity_mode": {
            "title": "Identity Mode",
            "type": "string"
          },
          "in_use": {
            "title": "In Use",
            "type": "boolean"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "readable": {
            "title": "Readable",
            "type": "boolean"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "root_path": {
            "title": "Root Path",
            "type": "string"
          },
          "visible": {
            "title": "Visible",
            "type": "boolean"
          },
          "writable": {
            "title": "Writable",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "label",
          "root_path",
          "readable",
          "writable",
          "identity_mode",
          "created_at",
          "visible",
          "in_use"
        ],
        "title": "StorageRootResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "BioPipeline2",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/storage/roots": {
      "get": {
        "description": "Every root, working or not.\n\nWithdrawn and unreachable roots are listed rather than filtered out: \"the\nshare is gone\" and \"nobody ever registered it\" are different problems and\na list that shows only working roots cannot tell them apart.",
        "operationId": "list_storage_roots_api_v1_storage_roots_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_StorageRootResponse_"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Storage Roots",
        "tags": [
          "storage"
        ]
      },
      "post": {
        "description": "Allowlist a path for task containers to read.\n\nThe path is refused if it is relative, if the platform cannot see it as a\ndirectory, if it overlaps the platform's own storage \u2014 which would give\nevery task a read-only view of every other run's outputs \u2014 or if it\noverlaps a root that already exists, whose permissions it would then\nsilently contradict. `RootRejected` is a `ValidationFailed`, so each of\nthose reaches the caller as a 422 saying which one it was.",
        "operationId": "register_api_v1_storage_roots_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRootRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageRootResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "storage"
        ]
      }
    },
    "/api/v1/storage/roots/{root_id}/reinstate": {
      "post": {
        "description": "Put a withdrawn root back, on a fresh attestation.\n\nNot an undo: the path is checked again and the attestation is made again,\nbecause the reason it was withdrawn may have been that the share's\npermissions changed.",
        "operationId": "reinstate_root_api_v1_storage_roots__root_id__reinstate_post",
        "parameters": [
          {
            "in": "path",
            "name": "root_id",
            "required": true,
            "schema": {
              "title": "Root Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReinstateRootRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageRootResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reinstate Root",
        "tags": [
          "storage"
        ]
      }
    },
    "/api/v1/storage/roots/{root_id}/revoke": {
      "post": {
        "description": "Stop offering a root.\n\nWithdrawn, not deleted: a delivery that went here still names it, and who\nattested it is part of the record whether or not it is still mounted.\nAnything already running keeps the mount it was given.",
        "operationId": "revoke_root_api_v1_storage_roots__root_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "root_id",
            "required": true,
            "schema": {
              "title": "Root Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeRootRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageRootResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Root",
        "tags": [
          "storage"
        ]
      }
    },
...
  }
}